+++ src/Twilio/Rest/Api/V2010/Account/CallList.php
@@ -22,19 +22,11 @@ use Twilio\Options;
 use Twilio\Stream;
 use Twilio\Values;
 use Twilio\Version;
-use Twilio\InstanceContext;
 use Twilio\Serialize;
-use Twilio\Rest\Api\V2010\Account\Call\FeedbackSummaryList;
 
 
-/**
- * @property FeedbackSummaryList $feedbackSummaries
- * @method \Twilio\Rest\Api\V2010\Account\Call\FeedbackSummaryContext feedbackSummaries(string $sid)
- */
 class CallList extends ListResource
     {
-    protected $_feedbackSummaries = null;
-
     /**
      * Construct the CallList
      *
@@ -286,55 +278,6 @@ class CallList extends ListResource
             $this->solution['accountSid'],
             $sid
         );
-    }
-
-    /**
-     * Access the feedbackSummaries
-     */
-    protected function getFeedbackSummaries(): FeedbackSummaryList
-    {
-        if (!$this->_feedbackSummaries) {
-            $this->_feedbackSummaries = new FeedbackSummaryList(
-                $this->version,
-                $this->solution['accountSid']
-            );
-        }
-        return $this->_feedbackSummaries;
-    }
-
-    /**
-     * Magic getter to lazy load subresources
-     *
-     * @param string $name Subresource to return
-     * @return \Twilio\ListResource The requested subresource
-     * @throws TwilioException For unknown subresources
-     */
-    public function __get(string $name)
-    {
-        if (\property_exists($this, '_' . $name)) {
-            $method = 'get' . \ucfirst($name);
-            return $this->$method();
-        }
-
-        throw new TwilioException('Unknown subresource ' . $name);
-    }
-
-    /**
-     * Magic caller to get resource contexts
-     *
-     * @param string $name Resource to return
-     * @param array $arguments Context parameters
-     * @return InstanceContext The requested resource context
-     * @throws TwilioException For unknown resource
-     */
-    public function __call(string $name, array $arguments): InstanceContext
-    {
-        $property = $this->$name;
-        if (\method_exists($property, 'getContext')) {
-            return \call_user_func_array(array($property, 'getContext'), $arguments);
-        }
-
-        throw new TwilioException('Resource does not have a context');
     }
 
     /**