+++ src/Twilio/Rest/Messaging/V1/TollfreeVerificationOptions.php
@@ -32,7 +32,7 @@ abstract class TollfreeVerificationOptions
      * @param string $businessContactFirstName The first name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactLastName The last name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactEmail The email address of the contact for the business or organization using the Tollfree number.
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @param string $externalReferenceId An optional external reference ID supplied by customer and echoed back on status retrieval.
      * @return CreateTollfreeVerificationOptions Options builder
      */
@@ -111,7 +111,7 @@ abstract class TollfreeVerificationOptions
      * @param string $businessContactFirstName The first name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactLastName The last name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactEmail The email address of the contact for the business or organization using the Tollfree number.
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @param string $editReason Describe why the verification is being edited. If the verification was rejected because of a technical issue, such as the website being down, and the issue has been resolved this parameter should be set to something similar to 'Website fixed'.
      * @return UpdateTollfreeVerificationOptions Options builder
      */
@@ -182,7 +182,7 @@ class CreateTollfreeVerificationOptions extends Options
      * @param string $businessContactFirstName The first name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactLastName The last name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactEmail The email address of the contact for the business or organization using the Tollfree number.
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @param string $externalReferenceId An optional external reference ID supplied by customer and echoed back on status retrieval.
      */
     public function __construct(
@@ -350,9 +350,9 @@ class CreateTollfreeVerificationOptions extends Options
     }
 
     /**
-     * The phone number of the contact for the business or organization using the Tollfree number.
+     * The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      *
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @return $this Fluent Builder
      */
     public function setBusinessContactPhone(string $businessContactPhone): self
@@ -461,7 +461,7 @@ class UpdateTollfreeVerificationOptions extends Options
      * @param string $businessContactFirstName The first name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactLastName The last name of the contact for the business or organization using the Tollfree number.
      * @param string $businessContactEmail The email address of the contact for the business or organization using the Tollfree number.
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @param string $editReason Describe why the verification is being edited. If the verification was rejected because of a technical issue, such as the website being down, and the issue has been resolved this parameter should be set to something similar to 'Website fixed'.
      */
     public function __construct(
@@ -739,9 +739,9 @@ class UpdateTollfreeVerificationOptions extends Options
     }
 
     /**
-     * The phone number of the contact for the business or organization using the Tollfree number.
+     * The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      *
-     * @param string $businessContactPhone The phone number of the contact for the business or organization using the Tollfree number.
+     * @param string $businessContactPhone The E.164 formatted phone number of the contact for the business or organization using the Tollfree number.
      * @return $this Fluent Builder
      */
     public function setBusinessContactPhone(string $businessContactPhone): self